requirements.txt
aniso8601==9.0.1
async-generator==1.10
attrs==23.1.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
Flask==2.2.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.1
python-dotenv==0.21.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
tomli==2.0.1
tqdm==4.64.0
trio==0.22.0
trio-websocket==0.10.2
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.4
wsproto==1.2.0
zipp==3.15.0